requirements.txt
numpy
pandas
scikit-learn
joblib
pyserial